Move the brushTarget null check
The brushTarget null check was happening after a condition that used brushTarget members.

Other instances of this in other files seem to forgo the null check entirely suggesting that it's not needed, but I decided to propose this change here at least in the interest of good practice. On a larger scale, I would consider moving the null check and the layer check to their own function since they seem to be used in multiple places.

I'm also not a fan of the magic number on the layer check, but I feel that is a separate, larger change and I worry that I'm starting to mess with things I don't fully understand (that said, it can't be a good idea to have this number in so many different places).